added access user control to pages related to staff and settings

+++ src/main/java/in/bloomington/incident/control/DamageTypeController.java
@@ -7,6 +7,7 @@ package in.bloomington.incident.control;
  */
 
 import java.util.List;
+import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,7 +21,9 @@ import javax.validation.Valid;
 // import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestParam;
 import in.bloomington.incident.service.DamageTypeService;
+import in.bloomington.incident.service.UserService;
 import in.bloomington.incident.model.DamageType;
+import in.bloomington.incident.model.User;
 
 
 @Controller
@@ -28,20 +31,36 @@ public class DamageTypeController extends TopController{
 
 		@Autowired
 		DamageTypeService damageTypeService;
+		@Autowired 
+    private HttpSession session;
+		@Autowired
+    UserService userService;
 		
 		@GetMapping("/damageTypes")
     public String getAll(Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}				
         model.addAttribute("types", damageTypeService.getAll());
         return "damageTypes";
     }
 		@GetMapping("/damageType/new")
     public String newDamageType(Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				DamageType damageType = new DamageType();
         model.addAttribute("type", damageType);
         return "damageTypeAdd";
     }     
     @PostMapping("/damageType/add")
     public String addDamageType(@Valid DamageType damageType, BindingResult result, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
         if (result.hasErrors()) {
             return "addDamageType";
         }
@@ -54,6 +73,10 @@ public class DamageTypeController extends TopController{
 
 		@GetMapping("/damageType/edit/{id}")
 		public String showEditForm(@PathVariable("id") int id, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				DamageType type = null;
 				try{
 						type = damageTypeService.findById(id);
@@ -71,8 +94,13 @@ public class DamageTypeController extends TopController{
 				return "dDamageTypeUpdate";
 		}
 		@PostMapping("/damageType/update/{id}")
-		public String updateDamageType(@PathVariable("id") int id, @Valid DamageType type, 
-														 BindingResult result, Model model) {
+		public String updateDamageType(@PathVariable("id") int id,
+																	 @Valid DamageType type, 
+																	 BindingResult result, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				if (result.hasErrors()) {
 						type.setId(id);
 						return "damageTypeUpdate";
@@ -86,7 +114,10 @@ public class DamageTypeController extends TopController{
 		
 		@GetMapping("/damageType/delete/{id}")
 		public String deleteDamageType(@PathVariable("id") int id, Model model) {
-
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				try{
 						DamageType type = damageTypeService.findById(id);
 						damageTypeService.delete(id);
@@ -104,5 +135,24 @@ public class DamageTypeController extends TopController{
 					 
 				return "damageTypes";
 		}
-		
+		private User findUserFromSession(HttpSession session){
+				User user = null;
+				User user2 = getUserFromSession(session);
+				if(user2 != null){
+						user = userService.findById(user2.getId());
+				}
+				return user;
+    }
+		private String canUserAccess(HttpSession session){
+				User user = findUserFromSession(session);
+				if(user == null){
+						return "redirect:/login";
+				}
+				if(!user.isAdmin()){
+						addMessage("you can not access");
+						addMessagesAndErrorsToSession(session);
+						return "redirect:staff";
+				}
+				return "";
+		}		
 }
