+++ src/main/java/in/bloomington/incident/control/PersonTypeController.java
@@ -7,6 +7,7 @@ package in.bloomington.incident.control;
  */
 
 import java.util.List;
+import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,7 +21,9 @@ import javax.validation.Valid;
 // import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestParam;
 import in.bloomington.incident.service.PersonTypeService;
+import in.bloomington.incident.service.UserService;
 import in.bloomington.incident.model.PersonType;
+import in.bloomington.incident.model.User;
 import in.bloomington.incident.utils.Helper;
 
 @Controller
@@ -28,20 +31,37 @@ public class PersonTypeController extends TopController{
 
 		@Autowired
 		PersonTypeService personTypeService;
+		@Autowired 
+    private HttpSession session;
+		@Autowired
+    UserService userService;
+		
 		
 		@GetMapping("/personTypes")
     public String getAll(Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
         model.addAttribute("types", personTypeService.getAll());
         return "personTypes";
     }
 		@GetMapping("/personType/new")
     public String newPersonType(Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				PersonType personType = new PersonType();
         model.addAttribute("type", personType);
         return "personTypeAdd";
     }     
     @PostMapping("/personType/add")
     public String addPersonType(@Valid PersonType personType, BindingResult result, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
         if (result.hasErrors()) {
             return "addPersonType";
         }
@@ -54,6 +74,10 @@ public class PersonTypeController extends TopController{
 
 		@GetMapping("/personType/edit/{id}")
 		public String showEditForm(@PathVariable("id") int id, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				PersonType type = null;
 				try{
 						type = personTypeService.findById(id);
@@ -71,8 +95,13 @@ public class PersonTypeController extends TopController{
 				return "personTypeUpdate";
 		}
 		@PostMapping("/personType/update/{id}")
-		public String updatePersonType(@PathVariable("id") int id, @Valid PersonType type, 
-														 BindingResult result, Model model) {
+		public String updatePersonType(@PathVariable("id") int id,
+																	 @Valid PersonType type, 
+																	 BindingResult result, Model model) {
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				if (result.hasErrors()) {
 						String error = Helper.extractErrors(result);
 						addError(error);
@@ -88,7 +117,10 @@ public class PersonTypeController extends TopController{
 		
 		@GetMapping("/personType/delete/{id}")
 		public String deletePersonType(@PathVariable("id") int id, Model model) {
-
+				String ret = canUserAccess(session);
+				if(!ret.isEmpty()){
+						return ret;
+				}
 				try{
 						PersonType type = personTypeService.findById(id);
 						personTypeService.delete(id);
@@ -106,5 +138,24 @@ public class PersonTypeController extends TopController{
 					 
 				return "personTypes";
 		}
-		
+		private User findUserFromSession(HttpSession session){
+				User user = null;
+				User user2 = getUserFromSession(session);
+				if(user2 != null){
+						user = userService.findById(user2.getId());
+				}
+				return user;
+    }		
+		private String canUserAccess(HttpSession session){
+				User user = findUserFromSession(session);
+				if(user == null){
+						return "redirect:/login";
+				}
+				if(!user.isAdmin()){
+						addMessage("you can not access");
+						addMessagesAndErrorsToSession(session);
+						return "redirect:staff";
+				}
+				return "";
+		}		
 }
